Remove: /mocks/data 를 제거하고 /nosql-data -> /data로 변경한다.
