--- src/mocks/data/categories.ts
@@ -1,158 +1,0 @@
-export const categories = [
-	{
-		id: 1,
-		name: '금융',
-	},
-	{
-		id: 2,
-		name: '비즈니스',
-	},
-	{
-		id: 3,
-		name: '쇼핑',
-	},
-	{
-		id: 4,
-		name: '음식',
-	},
-	{
-		id: 5,
-		name: '교육',
-	},
-	{
-		id: 6,
-		name: '건강 & 휘트니스',
-	},
-	{
-		id: 7,
-		name: '여행 & 교통',
-	},
-	{
-		id: 8,
-		name: '엔터테이먼트',
-	},
-	{
-		id: 9,
-		name: '인공지능',
-	},
-	{
-		id: 10,
-		name: 'CRM',
-	},
-	{
-		id: 11,
-		name: '라이프 스타일',
-	},
-	{
-		id: 12,
-		name: 'Crypto & Web3',
-	},
-	{
-		id: 13,
-		name: '생산',
-	},
-	{
-		id: 14,
-		name: '커뮤니티',
-	},
-	{
-		id: 15,
-		name: '개발 도구',
-	},
-	{
-		id: 16,
-		name: '출판',
-	},
-	{
-		id: 17,
-		name: '마케팅 & 광고',
-	},
-	{
-		id: 18,
-		name: '콘셉트 아트',
-	},
-	{
-		id: 19,
-		name: '패션',
-	},
-	{
-		id: 20,
-		name: '테크니컬',
-	},
-	{
-		id: 21,
-		name: '인포그래픽',
-	},
-	{
-		id: 22,
-		name: '패키징',
-	},
-	{
-		id: 23,
-		name: '캐리커쳐',
-	},
-	{
-		id: 24,
-		name: '3D그래픽',
-	},
-	{
-		id: 25,
-		name: '게임',
-	},
-	{
-		id: 26,
-		name: '인물',
-	},
-	{
-		id: 27,
-		name: '이벤트',
-	},
-	{
-		id: 28,
-		name: '웨딩',
-	},
-	{
-		id: 29,
-		name: '미술',
-	},
-	{
-		id: 30,
-		name: '건축',
-	},
-	{
-		id: 31,
-		name: '상업',
-	},
-	{
-		id: 32,
-		name: '사물',
-	},
-	{
-		id: 33,
-		name: '자연',
-	},
-	{
-		id: 34,
-		name: '음악',
-	},
-	{
-		id: 35,
-		name: '영화',
-	},
-	{
-		id: 36,
-		name: '다큐멘터리',
-	},
-	{
-		id: 37,
-		name: '요리',
-	},
-	{
-		id: 38,
-		name: '일상',
-	},
-	{
-		id: 39,
-		name: '편집',
-	},
-];
