--- src/mocks/data/reviews.ts
@@ -1,9 +1,0 @@
-export const reviews = [
-	{
-		id: 1,
-		commissionId: 1,
-		userId: 1,
-		createdAt: Date.now(),
-		content: '리뷰입니다~',
-	},
-];
