--- src/mocks/data/sections.ts
@@ -1,22 +1,0 @@
-export const sections = [
-	{
-		id: 1,
-		name: 'Android/iOS',
-	},
-	{
-		id: 2,
-		name: 'Web',
-	},
-	{
-		id: 3,
-		name: 'Illustration',
-	},
-	{
-		id: 4,
-		name: 'Photo',
-	},
-	{
-		id: 5,
-		name: 'Video',
-	},
-];
