--- src/mocks/data/tags.ts
@@ -1,86 +1,0 @@
-export const tags = [
-	{
-		id: 1,
-		sectionId: 1,
-		name: 'tag 1',
-		portfolioId: [1, 2, 12, 13, 23, 24],
-	},
-	{
-		id: 2,
-		sectionId: 1,
-		name: 'tag 2',
-		portfolioId: [1, 2, 3, 12, 13, 14, 23, 24, 25],
-	},
-	{
-		id: 3,
-		sectionId: 1,
-		name: 'tag 3',
-		portfolioId: [1, 2, 12, 13, 23, 24],
-	},
-	{
-		id: 4,
-		sectionId: 1,
-		name: 'tag 4',
-		portfolioId: [2, 3, 13, 14, 24, 25],
-	},
-	{
-		id: 5,
-		sectionId: 2,
-		name: 'tag 5',
-		portfolioId: [4, 5, 15, 16, 26, 27],
-	},
-	{
-		id: 6,
-		sectionId: 2,
-		name: 'tag 6',
-		portfolioId: [4, 15, 26],
-	},
-	{
-		id: 7,
-		sectionId: 2,
-		name: 'tag 7',
-		portfolioId: [4, 5, 15, 16, 26, 27],
-	},
-	{
-		id: 8,
-		sectionId: 2,
-		name: 'tag 8',
-		portfolioId: [5, 16, 27],
-	},
-	{
-		id: 9,
-		sectionId: 3,
-		name: 'tag 9',
-		portfolioId: [6, 7, 17, 18, 28, 29],
-	},
-	{
-		id: 10,
-		sectionId: 3,
-		name: 'tag 10',
-		portfolioId: [7, 18, 29],
-	},
-	{
-		id: 11,
-		sectionId: 4,
-		name: 'tag 11',
-		portfolioId: [8, 9, 19, 20, 30, 31],
-	},
-	{
-		id: 12,
-		sectionId: 4,
-		name: 'tag 12',
-		portfolioId: [9, 20, 31],
-	},
-	{
-		id: 13,
-		sectionId: 5,
-		name: 'tag 13',
-		portfolioId: [10, 11, 21, 22, 32, 33],
-	},
-	{
-		id: 14,
-		sectionId: 5,
-		name: 'tag 14',
-		portfolioId: [11, 22, 33],
-	},
-];
