+++ src/mocks/data/users.ts
@@ -1,131 +1,190 @@
-import { User } from "@/types";
+import { Users } from "@/types";
 
-export const users: User[] = [
-	{
-		id: 1,
-		nickname: '전문가 1',
-		profileImage: 'https://img1.daumcdn.net/thumb/R1280x0/?scode=mtistory2&fname=https%3A%2F%2Fblog.kakaocdn.net%2Fdn%2FnyjLl%2FbtsCr9rPmP3%2FW1k5kiFh3yLpkK6K1fkPJK%2Fimg.webp',
-		email: 'expert1@example.com',
-		name: '홍길동',
+export const users: Users = {
+	expert1: {
+		name: 'John',
+		email: 'john@example.com',
 		phone: '010-1234-1234',
+		nickname: 'John Lennon',
+		profileImage: 'https://img1.daumcdn.net/thumb/R1280x0/?scode=mtistory2&fname=https%3A%2F%2Fblog.kakaocdn.net%2Fdn%2FnyjLl%2FbtsCr9rPmP3%2FW1k5kiFh3yLpkK6K1fkPJK%2Fimg.webp',
 		authority: 'expert',
 		profile: {
-			introduce: '안녕하세요 전문가 1 입니다.',
-			location: '대구',
+			introduce: '안녕하세요 전문가 1 존 레논 입니다.',
+			location: '서울',
 			careers: ['네이버', '카카오'],
 			careerLength: '10년',
 			fields: ['Web', 'Frontend', 'Development'],
 			stacks: ['React', 'Vue', 'Node.js'],
-		},
-		activity: {
-			commissions: [ 1, 2, 3, 4 ],
-			score: 100,
+			score: 90,
 			contactTime: '언제나 가능',
 		},
-		likes: [1, 2, 3, 4, 5],
-		bookmarks: [1, 2, 3, 4, 5],
-		portfolios: [1, 4, 7, 10, 12, 15, 18, 21, 23, 26, 29, 32, 34, 37, 40, 43, 47, 50, 53, 56],
+		bookmarks: {},
+		likes: [],
 	},
-	{
-		id: 2,
-		nickname: '전문가 2',
-		profileImage: 'https://img1.daumcdn.net/thumb/R1280x0/?scode=mtistory2&fname=https%3A%2F%2Fblog.kakaocdn.net%2Fdn%2FnyjLl%2FbtsCr9rPmP3%2FW1k5kiFh3yLpkK6K1fkPJK%2Fimg.webp',
-		email: 'expert2@example.com',
-		name: '홍길동2',
+	expert2: {
+		name: 'Paul',
+		email: 'paul@example.com',
 		phone: '010-1234-1234',
+		nickname: 'Paul Mccartney',
+		profileImage: 'https://img1.daumcdn.net/thumb/R1280x0/?scode=mtistory2&fname=https%3A%2F%2Fblog.kakaocdn.net%2Fdn%2FnyjLl%2FbtsCr9rPmP3%2FW1k5kiFh3yLpkK6K1fkPJK%2Fimg.webp',
 		authority: 'expert',
 		profile: {
-			introduce: '안녕하세요 전문가 2 입니다.',
-			location: '대구',
+			introduce: '안녕하세요 전문가 2 폴 매카트니 입니다.',
+			location: '경기',
 			careers: ['네이버', '카카오'],
 			careerLength: '10년',
 			fields: ['Web', 'Frontend', 'Development'],
 			stacks: ['React', 'Vue', 'Node.js'],
-		},
-		activity: {
 			score: 100,
 			contactTime: '언제나 가능',
 		},
-		likes: [1, 2, 3, 4, 5],
-		bookmarks: [1, 2, 3, 4, 5],
-		portfolios: [2, 5, 8, 11, 13, 16, 19, 22, 24, 27, 30, 33, 35, 38, 41, 44, 48, 51, 54, 57,],
+		bookmarks: {},
+		likes: [],
 	},
-	{
-		id: 3,
-		nickname: '전문가 3',
-		profileImage: 'https://img1.daumcdn.net/thumb/R1280x0/?scode=mtistory2&fname=https%3A%2F%2Fblog.kakaocdn.net%2Fdn%2FnyjLl%2FbtsCr9rPmP3%2FW1k5kiFh3yLpkK6K1fkPJK%2Fimg.webp',
-		email: 'expert3@example.com',
-		name: '홍길동',
+	expert3: {
+		name: 'George',
+		email: 'george@example.com',
 		phone: '010-1234-1234',
+		nickname: 'George Harrison',
+		profileImage: 'https://img1.daumcdn.net/thumb/R1280x0/?scode=mtistory2&fname=https%3A%2F%2Fblog.kakaocdn.net%2Fdn%2FnyjLl%2FbtsCr9rPmP3%2FW1k5kiFh3yLpkK6K1fkPJK%2Fimg.webp',
 		authority: 'expert',
 		profile: {
-			introduce: '안녕하세요 전문가 3 입니다.',
-			location: '대구',
+			introduce: '안녕하세요 전문가 3 조지 해리슨 입니다.',
+			location: '강원도',
 			careers: ['네이버', '카카오'],
 			careerLength: '10년',
 			fields: ['Web', 'Frontend', 'Development'],
 			stacks: ['React', 'Vue', 'Node.js'],
-		},
-		activity: {
 			score: 100,
 			contactTime: '언제나 가능',
 		},
-		likes: [1, 2, 3, 4, 5],
-		bookmarks: [1, 2, 3, 4, 5],
-		portfolios: [3, 6, 9, 14, 17, 20, 25, 28, 31, 36, 39, 42, 45, 46, 49, 52, 55, 58, 59,],
+		bookmarks: {},
+		likes: [],
 	},
-	{
-		id: 100,
-		nickname: '의뢰인 1',
+	expert4: {
+		name: 'Ringo',
+		email: 'ringo@example.com',
+		phone: '010-1234-1234',
+		nickname: 'Ringo Starr',
 		profileImage: 'https://img1.daumcdn.net/thumb/R1280x0/?scode=mtistory2&fname=https%3A%2F%2Fblog.kakaocdn.net%2Fdn%2FnyjLl%2FbtsCr9rPmP3%2FW1k5kiFh3yLpkK6K1fkPJK%2Fimg.webp',
-		email: 'client1@example.com',
+		authority: 'expert',
+		profile: {
+			introduce: '안녕하세요 전문가 4 링고 스타 입니다.',
+			location: '경북',
+			careers: ['네이버', '카카오'],
+			careerLength: '10년',
+			fields: ['Web', 'Frontend', 'Development'],
+			stacks: ['React', 'Vue', 'Node.js'],
+			score: 100,
+			contactTime: '언제나 가능',
+		},
+		bookmarks: {},
+		likes: [],
+	},
+	expert5: {
+		name: 'Freddie',
+		email: 'freddie@example.com',
+		phone: '010-1234-1234',
+		nickname: 'John Lennon',
+		profileImage: 'https://img1.daumcdn.net/thumb/R1280x0/?scode=mtistory2&fname=https%3A%2F%2Fblog.kakaocdn.net%2Fdn%2FnyjLl%2FbtsCr9rPmP3%2FW1k5kiFh3yLpkK6K1fkPJK%2Fimg.webp',
+		authority: 'expert',
+		profile: {
+			introduce: '안녕하세요 전문가 5 프레디 머큐리 입니다.',
+			location: '제주',
+			careers: ['네이버', '카카오'],
+			careerLength: '10년',
+			fields: ['Web', 'Frontend', 'Development'],
+			stacks: ['React', 'Vue', 'Node.js'],
+			score: 100,
+			contactTime: '언제나 가능',
+		},
+		bookmarks: {},
+		likes: [],
+	},
+	client1: {
+		name: '김강철',
+		email: 'steel@example.com',
+		phone: '010-1234-1234',
+		nickname: '강철맨',
+		profileImage: 'https://img1.daumcdn.net/thumb/R1280x0/?scode=mtistory2&fname=https%3A%2F%2Fblog.kakaocdn.net%2Fdn%2FnyjLl%2FbtsCr9rPmP3%2FW1k5kiFh3yLpkK6K1fkPJK%2Fimg.webp',
 		authority: 'client',
 		profile: {
 			introduce: '안녕하세요 의뢰인 1 입니다.',
 			location: '대구',
-		},
-		activity: {
-			commissions: [ 1 ],
 			contactTime: '언제나 가능',
 		},
-		likes: [1, 2, 3, 4, 5],
-		bookmarks: [1, 2, 3, 4, 5],
-		portfolios: [],
+		bookmarks: {
+			portfolio1: {
+				title: 'test title',
+				summary: 'test summary',
+				thumbnailUrl: '',
+			},
+			portfolio2: {
+				title: 'test title',
+				summary: 'test summary',
+				thumbnailUrl: '',
+			},
+		},
+		likes: [],
 	},
-	{
-		id: 200,
-		nickname: '의뢰인 2',
+	client2: {
+		name: '김우주',
+		email: 'universe@example.com',
+		phone: '010-1234-1234',
+		nickname: '코스모스',
 		profileImage: 'https://img1.daumcdn.net/thumb/R1280x0/?scode=mtistory2&fname=https%3A%2F%2Fblog.kakaocdn.net%2Fdn%2FnyjLl%2FbtsCr9rPmP3%2FW1k5kiFh3yLpkK6K1fkPJK%2Fimg.webp',
-		email: 'client2@example.com',
 		authority: 'client',
 		profile: {
 			introduce: '안녕하세요 의뢰인 2 입니다.',
 			location: '대구',
-		},
-		activity: {
-			commissions: [ 2, 3 ],
 			contactTime: '언제나 가능',
 		},
-		likes: [1, 2, 3, 4, 5],
-		bookmarks: [1, 2, 3, 4, 5],
-		portfolios: [],
+		bookmarks: {},
+		likes: [],
 	},
-	{
-		id: 300,
-		nickname: '의뢰인 3',
+	client3: {
+		name: '스폰지밥',
+		email: 'underthesea@example.com',
+		phone: '010-1234-1234',
+		nickname: '스폰지밥',
 		profileImage: 'https://img1.daumcdn.net/thumb/R1280x0/?scode=mtistory2&fname=https%3A%2F%2Fblog.kakaocdn.net%2Fdn%2FnyjLl%2FbtsCr9rPmP3%2FW1k5kiFh3yLpkK6K1fkPJK%2Fimg.webp',
-		email: 'client3@example.com',
 		authority: 'client',
 		profile: {
 			introduce: '안녕하세요 의뢰인 3 입니다.',
 			location: '대구',
-		},
-		activity: {
-			commissions: [ 4 ],
 			contactTime: '언제나 가능',
 		},
-		likes: [1, 2, 3, 4, 5],
-		bookmarks: [1, 2, 3, 4, 5],
-		portfolios: [],
+		bookmarks: {},
+		likes: [],
 	},
-];
+	client4: {
+		name: '징징이',
+		email: 'squeed@example.com',
+		phone: '010-1234-1234',
+		nickname: '징징이',
+		profileImage: 'https://img1.daumcdn.net/thumb/R1280x0/?scode=mtistory2&fname=https%3A%2F%2Fblog.kakaocdn.net%2Fdn%2FnyjLl%2FbtsCr9rPmP3%2FW1k5kiFh3yLpkK6K1fkPJK%2Fimg.webp',
+		authority: 'client',
+		profile: {
+			introduce: '안녕하세요 의뢰인 4 입니다.',
+			location: '대구',
+			contactTime: '언제나 가능',
+		},
+		bookmarks: {},
+		likes: [],
+	},
+	client5: {
+		name: '집게사장',
+		email: 'crab@example.com',
+		phone: '010-1234-1234',
+		nickname: '집게사장',
+		profileImage: 'https://img1.daumcdn.net/thumb/R1280x0/?scode=mtistory2&fname=https%3A%2F%2Fblog.kakaocdn.net%2Fdn%2FnyjLl%2FbtsCr9rPmP3%2FW1k5kiFh3yLpkK6K1fkPJK%2Fimg.webp',
+		authority: 'client',
+		profile: {
+			introduce: '안녕하세요 의뢰인 5 입니다.',
+			location: '대구',
+			contactTime: '언제나 가능',
+		},
+		bookmarks: {},
+		likes: [],
+	},
+};
