+++ src/mocks/handlers/user.ts
@@ -1,8 +1,8 @@
 import { HttpResponse, http } from 'msw';
 
 import { LOGIN_ID } from '@/mocks/handlers';
-import { portfolios } from '@/mocks/nosql-data/portfolios';
-import { users } from '@/mocks/nosql-data/users';
+import { portfolios } from '@/mocks/data/portfolios';
+import { users } from '@/mocks/data/users';
 import { Commission, Portfolio, User } from '@/types';
 
 export const userHandlers= [
